rename/move files containing GameState member functions
