+++ config/jpn/arm9/delinks.txt
@@ -5,11 +5,11 @@
     .data       start:0x020eeb40 end:0x020f2e60 kind:data align:32
     .bss        start:0x020f2e60 end:0x02154e60 kind:bss align:32
 
-src/Combat/Main/GetBattleList.cpp:
+src/GameState/GameStateInstance.cpp:
 	complete
 	.text start:0x0200f25c end:0x0200f268
 
-src/Combat/Main/GetCombatantFromList.cpp:
+src/GameState/GameStateObjects.cpp:
 	complete
 	.text start:0x0200fbcc end:0x0200fd00
