+++ config/usa/arm9/delinks.txt
@@ -5,11 +5,11 @@
     .data       start:0x020eebe0 end:0x020f2e60 kind:data align:32
     .bss        start:0x020f2e60 end:0x021536e0 kind:bss align:32
 
-src/Combat/Main/GetBattleList.cpp:
+src/GameState/GameStateInstance.cpp:
 	complete
 	.text start:0x0200f398 end:0x0200f3a4
 
-src/Combat/Main/GetCombatantFromList.cpp:
+src/GameState/GameStateObjects.cpp:
 	complete
 	.text start:0x0200fd70 end:0x0200fea4
