+++ src/Combat/Overlay_0/GetCombatantByID.cpp
@@ -1,6 +1,6 @@
 #include <globaldefs.h>
 #include "Combat/Main/BattleList.h"
-#include "GameState.h"
+#include "GameState/GameState.h"
 
 struct GameObject* GetCombatantByID(int unused, int id) {
 	return GameState::GetInstance()->GetCombatantByIndex(id);
