+++ src/Graphics/LightingManager.cpp
@@ -2,7 +2,7 @@
 #include "World/Zone3D.h"
 #include "System/Memory.h"
 #include "System/Graphics.h"
-#include "GameState.h"
+#include "GameState/GameState.h"
 #include "Resource/GameResources.h"
 
 #if defined(jpn)
