+++ src/Grotto/Main/ActiveGrottoClass.cpp
@@ -1,7 +1,7 @@
 #include "Grotto/Main/ActiveGrottoClass.h"
 #include "Combat/Main/BattleList.h"
 #include "Grotto/Main/TreasureMapDataStructs.h"
-#include "GameState.h"
+#include "GameState/GameState.h"
 #include <globaldefs.h>
 
 #ifdef jpn
