+++ src/Grotto/Main/TreasureMapDataStructs.cpp
@@ -1,7 +1,7 @@
 #include "Grotto/Main/TreasureMapDataStructs.h"
 #include "std_library_functions.h"
 #include "Grotto/Main/GrottoStruct.h"
-#include "GameState.h"
+#include "GameState/GameState.h"
 #include "System/Memory.h"
 #include "Resource/GameResources.h"
