+++ src/Grotto/Main/TreasureMapMetadata.cpp
@@ -1,8 +1,7 @@
 #include "Grotto/Main/TreasureMapDataStructs.h"
 #include "System/Memory.h"
 #include "std_library_functions.h"
-#include "GameState.h"
-#include "Grotto/Main/GrottoStruct.h"
+#include "GameState/GameState.h"
 #include <globaldefs.h>
 
 #ifdef jpn
